+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 All notable changes to this project will be documented in this file. (Pending approval) This project adheres to [Semantic Versioning](http://semver.org/). You can read more on this at [keep a change log](http://keepachangelog.com/)
 
 # [Unreleased]
+- `Added` Config files to Service